Prevent firing of event handler
When calling the click() method, any event handler that might have been
bound to the element could be fired, which doesn't necessarily have the
expected behaviour. The finalize() method is meant to be called within
the containing quiz element at render time, meaning that all of the
quiz's properties might not yet be properly set up at the time it is
called. If the finalize() method of Question calls the click() method on
an input element, and the element has a handler which contains calls to
unset properties of the quiz element, an error could be thrown. On
chrome, when the click() method is called, the handler is called
asynchronously, and everything runs fine. But on firefox, the handler
runs kinda like synchronously, because the values read in the handler
are those present when the click() method was called not when the
handler was called.

Since the finalize() method is expected to be called with a
metadata only during render time, the checked attribute should be
manually set and event handlers should not be triggered.